+++ Lib/fontbakery/checks/tables.py
@@ -128,6 +128,12 @@ def check_required_tables(ttFont, config, is_variable_font):
 def check_unwanted_tables(ttFont):
     """Are there unwanted tables?"""
     UNWANTED_TABLES = {
+        "DSIG": (
+            "This font has a digital signature (DSIG table) which is only required"
+            " - even if only a placeholder - on old programs like MS Office 2013"
+            " in order to work properly.\n"
+            "The current recommendation is to completely remove the DSIG table."
+        ),
         "FFTM": "Table contains redundant FontForge timestamp info",
         "TTFA": "Redundant TTFAutohint table",
         "TSI0": "Table contains data only used in VTT",