+++ CHANGELOG.md
@@ -2,7 +2,9 @@ Below are the noteworthy changes from each release.
 A more detailed list of changes is available in the corresponding milestones for each release in the Github issue tracker (https://github.com/googlefonts/fontbakery/milestones?state=closed).
 
 ##  Upcoming release: 0.13.0 (a4?) (2024-Nov-??)
-  - ...
+### Deprecated checks
+#### Removed from the OpenType profile
+  - **DEPRECATED - [opentype/dsig]:** Merged into **[unwanted_tables]** on the `Universal` profile. (issue #4865)
 
 
 ## 0.13.0a3 (2024-Oct-25)